añadido el nuevo componente HeroSection y eliminado los componentes obsoletos de SectionHero
